Pay: bills become a full-width table with a detail drawer
Cards suited the top metrics but a list of bills is tabular data — it reads better
as rows with columns you can scan down. Replaces the side-by-side master/detail (and
its stack of bordered cards) with:

- BillTable: one bordered container, hairline-divided rows, columns Bill · Due ·
  Amount · Earns · Status · ›. Period toggle (week/month/quarter) and add live in the
  toolbar. Secondary columns drop out at narrow widths so it collapses to Bill ·
  Amount · Status on a phone. Still an agenda — ordered by due, recurrence
  consolidated to "×N", nothing repeated.
- BillManager: selecting a row (or adding) opens the detail in the responsive Sheet —
  a right drawer on desktop, a bottom sheet with pull-handle on mobile.
- BillActivity (coming up / recently paid) moves under the table as its own module,
  since the detail is no longer a persistent pane.

Removes BillWorkspace and BillList (superseded). BillDetailPane flows naturally in
the scrollable drawer (dropped h-full).
